Improve modularity by collect one-to-five liner require files into a single file.
Easier to get a good overview of what's going on when doing it this way. No point having one-liner .els just to require a single mode. Still not sure how to best organize things, but going to try changing namespace prefix to something more personal for the time, and chuck them al into -modules.el until that gets too big.