Show some line number information in frame-title
Since mode-line-position doesn't work in frame-title-format (%c and %l
are ignored), we use (:eval) to get the line-number-at-pos
manually. This has the limitation of not updating on every navigation,
but updates on C-l so it's a decent compromise (at the moment).

Also remove some pointless flattening of the mode-line (since it's not
being displayed anyway).